sort declarations
