+++ sample.h
@@ -45,77 +45,6 @@ namespace EPP
         Value() : binary(0){};
     };
 
-    /**
-     * Templates depending on the clients data model
-     */
-    template <class ClientSample>
-    class SampleSubset : public Subset, protected Blob
-    {
-        friend class SubsetStream;
-
-    public:
-        Event events;
-        Measurement X, Y;
-        Polygon polygon;
-        const SampleSubset *const parent;
-        std::vector<const SampleSubset *> children;
-
-        SampleSubset(
-            const ClientSample &sample)
-            : Subset(sample, true), parent(nullptr), events(sample.events)
-        {
-            for (Event event = 0; event < sample.events; event++)
-                for (Measurement measurement = 0; measurement < sample.measurements; measurement++)
-                    if (sample(event, measurement) < 0 || sample(event, measurement) > 1)
-                    {
-                        member(event, false);
-                        --events;
-                    };
-        };
-
-        json tree() const noexcept
-        {
-            static int subset_count = 0;
-            json subset;
-            subset["ID"] = ++subset_count;
-            if (parent)
-            {
-                subset["X"] = X;
-                subset["Y"] = Y;
-
-                json polygon;
-                for (auto &point : this->polygon)
-                {
-                    json vertex;
-                    vertex[0] = point.x();
-                    vertex[1] = point.y();
-                    polygon += vertex;
-                };
-                subset["polygon"] = polygon;
-            }
-            subset["events"] = events;
-
-            if (this->children.size() > 0)
-            {
-                json children;
-                for (const SampleSubset *child : this->children)
-                    children += child->tree();
-                subset["children"] = children;
-            }
-            return subset;
-        };
-
-        SampleSubset(
-            const ClientSample &sample,
-            const SampleSubset *parent,
-            const Subset &subset) : Subset(sample, subset), parent(parent){};
-
-        ~SampleSubset()
-        {
-            for (auto &child : children)
-                delete child;
-        };
-    };
 
     template <typename _float>
     class DefaultSample : public Sample
